requirements.txt
altgraph==0.17
attrs==21.2.0
et-xmlfile==1.1.0
iniconfig==1.1.1
macholib==1.14
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pytest==6.2.4
toml==0.10.2
